+++ tests/__init__.py
@@ -5,4 +5,3 @@ Created on Thu Jul 12 11:41:00 2018
 
 @author: Paolo Cozzi <cozzi@ibba.cnr.it>
 """
-